# docs/swagger.yaml
info:
  contact: {}
paths:
  /get.add_user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: re_password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code","message"}
          schema:
            type: obj
      summary: 新增用户
      tags:
      - 用户模块
  /get.del_user:
    post:
      consumes:
      - application/json
      parameters:
      - description: "1"
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code","message"}
          schema:
            type: obj
      summary: 删除用户
      tags:
      - 用户模块
  /get.user_info:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: obj
      tags:
      - 获取用户详情
  /get.user_list:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: obj
      tags:
      - 获取用户列表
  /update_user:
    post:
      consumes:
      - application/json
      parameters:
      - description: "1"
        in: formData
        name: id
        type: string
      - description: taibao
        in: formData
        name: name
        type: string
      - description: "123123"
        in: formData
        name: password
        type: string
      - description: "123123"
        in: formData
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code","message"}
          schema:
            type: obj
      summary: 更新用户
      tags:
      - 用户模块
swagger: "2.0"